fix(web): change lpaq file grouping
