+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/community-infrastructure-levy/has-community-infrastructure-levy/has-community-infrastructure-levy.controller.js
@@ -1,16 +1,16 @@
-import { dayMonthYearHourMinuteToISOString } from '#lib/dates.js';
+import { convertFromYesNoNullToBooleanOrNull } from '#lib/boolean-formatter.js';
 import logger from '#lib/logger.js';
 import { addNotificationBannerToSession } from '#lib/session-utilities.js';
 import { getOriginPathname, isInternalUrl } from '#lib/url-utilities.js';
-import { getLpaQuestionnaireFromId } from '../lpa-questionnaire.service.js';
-import * as mapper from './infrastructure-levy-expected-date.mapper.js';
-import * as service from './infrastructure-levy-expected-date.service.js';
+import { getLpaQuestionnaireFromId } from '../../lpa-questionnaire.service.js';
+import * as mapper from './has-community-infrastructure-levy.mapper.js';
+import * as service from './has-community-infrastructure-levy.service.js';
 
 /**
  * @param {import('@pins/express/types/express.js').Request} request
  * @param {import('@pins/express/types/express.js').RenderedResponse<any, any, Number>} response
  */
-const renderChangeInfrastructureLevyExpectedDate = async (request, response) => {
+const renderChangeHasCommunityInfrastructureLevy = async (request, response) => {
 	try {
 		const { currentAppeal, session, errors, originalUrl, apiClient } = request;
 		const origin = originalUrl.split('/').slice(0, -2).join('/');
@@ -20,11 +20,10 @@ const renderChangeInfrastructureLevyExpectedDate = async (request, response) => 
 			currentAppeal.lpaQuestionnaireId
 		);
 
-		const mappedPageContents = mapper.changeInfrastructureLevyExpectedDate(
+		const mappedPageContents = mapper.changeHasCommunityInfrastructureLevy(
 			currentAppeal,
-			dayMonthYearHourMinuteToISOString(session.infrastructureLevyExpectedDate) ||
-				lpaQuestionnaireData.infrastructureLevyExpectedDate ||
-				null,
+			convertFromYesNoNullToBooleanOrNull(session.hasCommunityInfrastructureLevy) ??
+				lpaQuestionnaireData.hasInfrastructureLevy,
 			origin
 		);
 
@@ -43,13 +42,13 @@ const renderChangeInfrastructureLevyExpectedDate = async (request, response) => 
  * @param {import('@pins/express/types/express.js').Request} request
  * @param {import('@pins/express/types/express.js').RenderedResponse<any, any, Number>} response
  */
-export const getChangeInfrastructureLevyExpectedDate = renderChangeInfrastructureLevyExpectedDate;
+export const getChangeHasCommunityInfrastructureLevy = renderChangeHasCommunityInfrastructureLevy;
 
 /**
  * @param {import('@pins/express/types/express.js').Request} request
  * @param {import('@pins/express/types/express.js').RenderedResponse<any, any, Number>} response
  */
-export const postChangeInfrastructureLevyExpectedDate = async (request, response) => {
+export const postChangeHasCommunityInfrastructureLevy = async (request, response) => {
 	try {
 		const {
 			apiClient,
@@ -60,15 +59,10 @@ export const postChangeInfrastructureLevyExpectedDate = async (request, response
 		} = request;
 
 		if (errors) {
-			return renderChangeInfrastructureLevyExpectedDate(request, response);
+			return renderChangeHasCommunityInfrastructureLevy(request, response);
 		}
 
-		/** @type {import('#lib/dates.js').DayMonthYearHourMinute} */
-		session.infrastructureLevyExpectedDate = {
-			day: request.body['levy-expected-date-day'],
-			month: request.body['levy-expected-date-month'],
-			year: request.body['levy-expected-date-year']
-		};
+		session.hasCommunityInfrastructureLevy = request.body['hasCommunityInfrastructureLevyRadio'];
 
 		const currentUrl = getOriginPathname(request);
 		const origin = currentUrl.split('/').slice(0, -2).join('/');
@@ -79,21 +73,21 @@ export const postChangeInfrastructureLevyExpectedDate = async (request, response
 			});
 		}
 
-		await service.changeInfrastructureLevyExpectedDate(
+		await service.changeHasCommunityInfrastructureLevy(
 			apiClient,
 			appealId,
 			currentAppeal.lpaQuestionnaireId,
-			session.infrastructureLevyExpectedDate
+			session.hasCommunityInfrastructureLevy
 		);
 
 		addNotificationBannerToSession({
 			session,
 			bannerDefinitionKey: 'changePage',
 			appealId,
-			text: 'Expected levy adoption date changed'
+			text: 'Community infrastructure levy status changed'
 		});
 
-		delete request.session.infrastructureLevyExpectedDate;
+		delete request.session.hasCommunityInfrastructureLevy;
 
 		if (!origin.startsWith('/')) {
 			throw new Error('unexpected originalUrl');
@@ -104,7 +98,7 @@ export const postChangeInfrastructureLevyExpectedDate = async (request, response
 		logger.error(error);
 	}
 
-	delete request.session.infrastructureLevyExpectedDate;
+	delete request.session.hasCommunityInfrastructureLevy;
 
 	return response.status(500).render('app/500.njk');
 };
