+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/community-infrastructure-levy/infrastructure-levy-adopted-date/infrastructure-levy-adopted-date.mapper.js
@@ -1,5 +1,5 @@
 /**
- * @typedef {import('../../appeal-details.types.js').WebAppeal} Appeal
+ * @typedef {import('../../../appeal-details.types.js').WebAppeal} Appeal
  */
 import { appealShortReference } from '#lib/appeals-formatter.js';
 import { dateInput } from '#lib/mappers/components/page-components/date.js';
