+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/environmental-impact-assessment/eia-environmental-impact-schedule/eia-environmental-impact-schedule.controller.js
@@ -4,7 +4,7 @@ import {
 	mapEiaEnvironmentalImpactScheduleRadioValueForService
 } from './eia-environmental-impact-schedule.mapper.js';
 import { changeEiaEnvironmentalImpactSchedule } from './eia-environmental-impact-schedule.service.js';
-import { getLpaQuestionnaireFromId } from '../lpa-questionnaire.service.js';
+import { getLpaQuestionnaireFromId } from '../../lpa-questionnaire.service.js';
 import { addNotificationBannerToSession } from '#lib/session-utilities.js';
 
 /**
