+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/lpa-questionnaire.router.js
@@ -27,12 +27,12 @@ import hasProtectedSpeciesRouter from './has-protected-species/has-protected-spe
 import affectsScheduledMonumentRouter from './affects-scheduled-monument/affects-scheduled-monument.router.js';
 import isGypsyOrTravellerSiteRouter from './is-gypsy-or-traveller-site/is-gypsy-or-traveller-site.router.js';
 import isAonbNationalLandscapeRouter from './is-aonb-national-landscape/is-aonb-national-landscape.router.js';
-import hasCommunityInfrastructureLevyRouter from './has-community-infrastructure-levy/has-community-infrastructure-levy.router.js';
-import isInfrastructureLevyFormallyAdoptedRouter from './is-infrastructure-levy-formally-adopted/is-infrastructure-levy-formally-adopted.router.js';
-import infrastructureLevyAdoptedDateRouter from './infrastructure-levy-adopted-date/infrastructure-levy-adopted-date.router.js';
-import infrastructureLevyExpectedDateRouter from './infrastructure-levy-expected-date/infrastructure-levy-expected-date.router.js';
-import eiaEnvironmentalImpactScheduleRouter from './eia-environmental-impact-schedule/eia-environmental-impact-schedule.router.js';
-import eiaDevelopmentDescriptionRouter from './eia-development-description/eia-development-description.router.js';
+import hasCommunityInfrastructureLevyRouter from './community-infrastructure-levy/has-community-infrastructure-levy/has-community-infrastructure-levy.router.js';
+import isInfrastructureLevyFormallyAdoptedRouter from './community-infrastructure-levy/is-infrastructure-levy-formally-adopted/is-infrastructure-levy-formally-adopted.router.js';
+import infrastructureLevyAdoptedDateRouter from './community-infrastructure-levy/infrastructure-levy-adopted-date/infrastructure-levy-adopted-date.router.js';
+import infrastructureLevyExpectedDateRouter from './community-infrastructure-levy/infrastructure-levy-expected-date/infrastructure-levy-expected-date.router.js';
+import eiaEnvironmentalImpactScheduleRouter from './environmental-impact-assessment/eia-environmental-impact-schedule/eia-environmental-impact-schedule.router.js';
+import eiaDevelopmentDescriptionRouter from './environmental-impact-assessment/eia-development-description/eia-development-description.router.js';
 import procedurePreferenceRouter from './procedure-preference/procedure-preference.router.js';
 import neighbouringSiteAccessRouter from './neighbouring-site-access/neighbouring-site-access.router.js';
 import designatedSitesRouter from './designated-sites/designated-sites.router.js';
