+++ appeals/web/src/server/lib/mappers/data/lpa-questionnaire/submappers/map-eia-development-description.js
@@ -1,4 +1,4 @@
-import { eiaDescriptions } from '#appeals/appeal-details/lpa-questionnaire/eia-development-description/eia-development-description.mapper.js';
+import { eiaDescriptions } from '#appeals/appeal-details/lpa-questionnaire/environmental-impact-assessment/eia-development-description/eia-development-description.mapper.js';
 import { textSummaryListItem } from '#lib/mappers/index.js';
 
 /** @type {import("../mapper.js").SubMapper} */
